feat(spendPoints): available points calc bug fix
As stated in commit aeb911902d265a45edeb1362924722218cc100c3,
the overall points available per payer was not taken into account when
traversing through transactions in least to most order and figuring out
the spending metrics.

This commit fixes that issue by figuring out the total availble points
for each payer and flagging whether the maximum points availble by a
certain payer has been reached or passed yet. This enables us to not
spend more points than are available under a payer and keep on
continuing until we have met our points requirement.

I was having trouble with the calculation encountering the exit condition
of meeting the points quota since I was not calculating the points
remaining to be spent properly. I have fixed that here as well.