feat(tests): update test framework with lodash
I updated the way I was testing the spendPoints function and am now
using lodash to compare the actual and expected results to see whether
the tests are passing or not. This way also allows me to add tests much
easier and maintain them better than before.

I renamed test.ts to reflect that this file is testing for only the
spendPoints function. This way I can add more tests for other files and
functions later.