feat(): almost finished
I updated transaction storage to also include total available points
and points per payer.  For this I created a new interface called
TransactionMetrics.  This moves the operation of calculating the total
points, something that was done previously by the `spendPoints`
function, into the main workflow and is kept in memory for the entire
lifetime of the application and can be accessed by any of the other
files.  This means that now every time a new transaction is created the
`transactionMetrics` need to be updated as well. To aid this function, I
separated it out into its own file called `addTransaction`.

There's also the last required endpoint for checking the current balance
that was added to index.ts. This is also aided by TransactionMetrics
since the totalPerPayer field records the relevant metric. I just needed
to change the format of the output according to the guidelines.

For spendPoints specifically, I tried simplifying the logic. Previously
there were way too many variables trying to keep track of the total
points that was spent. Now, it's a little easier to follow.